requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.1.4

# Testing
#...

# Code quality
flake8==3.6.0
